rename icon-interface to icon-type
