+++ src/components/fundamentals/icons/bell-icon.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import IconProps from "./icon-interface"
+import IconProps from "./icon-type"
 
 const BellIcon: React.FC<IconProps> = ({
   size = "24px",
